+++ Shop/StatsUpdater.cs
@@ -15,7 +15,7 @@ namespace Shop
             FilePath = filePath;
         }
 
-        public virtual void Update(Dictionary<string, int> cart, string name, int money)
+        public virtual void Update(Dictionary<string, int> cart, string name, double money)
         {
 
         }
